Responded to Menghan and Easwar's comments
